juegos_gramaticales/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
SQLAlchemy==2.0.20
alembic==1.11.1
email-validator==2.0.0
python-dotenv==1.0.0
pytest==7.4.0
pytest-flask==1.2.0
black==23.7.0
flake8==6.1.0
pylint==2.17.4
gunicorn==21.2.0
psycopg2-binary==2.9.7
python-dateutil==2.8.2
blinker==1.6.2
click==8.1.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
